refactor: merge write stream management
